Структура Cond вынесена в отдельный модуль cond
